Revert "[Death Knight] Allow rune count expressions to represent partial runes."
This reverts commit bc2c465c7e10b284cdd2598903c1c005542fc798.

+++ engine/class_modules/sc_death_knight.cpp
@@ -535,7 +535,7 @@ public:
   void      trigger_t17_4pc_frost( const action_state_t* );
   void      trigger_t17_4pc_unholy( const action_state_t* );
   void      apply_diseases( action_state_t* state, unsigned diseases );
-  double    runes_count( rune_type rt, bool include_death, int position );
+  int       runes_count( rune_type rt, bool include_death, int position );
   double    runes_cooldown_any( rune_type rt, bool include_death, int position );
   double    runes_cooldown_all( rune_type rt, bool include_death, int position );
   double    runes_cooldown_time( rune_t* r );
@@ -7655,7 +7655,7 @@ void death_knight_t::apply_diseases( action_state_t* state, unsigned diseases )
 
 // death_knight_t::runes_count ==============================================
 // how many runes of type rt are available
-double death_knight_t::runes_count( rune_type rt, bool include_death, int position )
+int death_knight_t::runes_count( rune_type rt, bool include_death, int position )
 {
   double result = 0;
   // positional checks first
@@ -7686,7 +7686,7 @@ double death_knight_t::runes_count( rune_type rt, bool include_death, int positi
       }
     }
   }
-  return result;
+  return static_cast<int>( result );
 }
 
 // death_knight_t::runes_cooldown_any =======================================
